move filelock test
