+++ tests/test_singularity_permanent_cache.py
@@ -21,10 +21,11 @@
 from pathlib import Path
 import sys
 import tempfile
-
+import threading
+import time
 import pytest
 
-from singularity_permanent_cache import get_cache_dir_from_env, uri_to_filename, main
+from singularity_permanent_cache import get_cache_dir_from_env, uri_to_filename, main, SimpleUnixFileLock
 
 # CACHE DIR TESTS
 
@@ -51,7 +52,7 @@ def test_get_cache_dir_from_env_no_env(monkeypatch):
     error.match("Please set 'SINGULARITY_PERMANENTCACHEDIR' or "
                 "'SINGULARITY_CACHEDIR.")
 
-# URI tests
+
 URIS = [
     ("docker://quay.io/biocontainers/bedtools:2.23.0--hdbcaa40_3",
      "docker_quay.io_biocontainers_bedtools_2.23.0--hdbcaa40_3"),
@@ -72,6 +73,39 @@ URIS = [
 @pytest.mark.parametrize(["uri", "result"], URIS)
 def test_uri_to_filename(uri, result):
     assert uri_to_filename(uri) == result
+
+
+def test_filelock():
+    count = 5
+    lockfile = tempfile.mktemp()
+    _, execution_times_file = tempfile.mkstemp()
+    Path(execution_times_file).write_text("")  # make sure the file exists
+
+    def filelock_test():
+        with SimpleUnixFileLock(lockfile):
+            # We wait 0.1 second to make sure current time is distinct if the
+            # lock works.
+            time.sleep(0.1)
+            current_time = str(round(time.time(), 1))
+            # Use append mode here
+            with open(execution_times_file, mode="at") as file_h:
+                file_h.write(current_time + "\n")
+
+    # Start all the processes in quick succession.
+    threads = [threading.Thread(target=filelock_test) for _ in range(count)]
+
+    for thread in threads:
+        thread.start()
+
+    for thread in threads:
+        thread.join()
+
+    # Get the epoch time (in seconds) from the file. If the lock does not work
+    # We expect multiple processes to have executed at the same time.
+    # We use a set comprehension to collapse times that are the same.
+    execution_times = {float(line) for line in
+                       Path(execution_times_file).read_text().splitlines()}
+    assert len(execution_times) == count
